fix(test): isolate count_tokens breaker regressions in unit-only file (no-web-impact)
The two new regressions added in this PR's first commit referenced
`anthropicUpstreamErrorCounterCacheStub` + `rateLimitAccountRepoStub`,
which are declared in `ratelimit_service_401_test.go` under
`//go:build unit`. The host file
`gateway_anthropic_apikey_passthrough_test.go` is intentionally
untagged so its tests run under both unit + integration; placing the
new tests there made `go test -tags=integration ./...` fail to
compile with `undefined: anthropicUpstreamErrorCounterCacheStub`
(observed on PR #280 CI test-integration + golangci-lint jobs).

Move the two regressions to a new
`gateway_count_tokens_breaker_test.go` with `//go:build unit` so they
live alongside the stubs they depend on, while the existing untagged
file keeps running under both tag sets unchanged. Net behavior
identical: the same two regressions still execute under
`go test -tags=unit ./...`; the integration tag no longer tries to
compile them.

Verified locally: `go test -tags=unit -run 'TestGatewayService_ForwardCountTokens_(400|429)' ./internal/service/` PASS;
`go vet -tags=integration ./internal/service/` clean;
`go test -tags=integration -count=1 -run '^TestNothingMatchesThis$' ./internal/service/` builds.
